Think I figured out Timing
The problem is if there is a collission with the opcodes being a part of a delay,
you will accidentally only grab 2 bytes instead of 3.
The look ahead is flawed. I am assuming just processing each command sequentially with only one byte look ahead would remove this issue as we
wouldn't have to worry about this special case based on the logic on
miditone's page.

Showing my debugging for historical purposes