Fixed Timing Logic
We unfortnuately cant do any clean generic looping look ahead as we can't assume anything isn't the other part of a delay number which is how we started losing notes and order.

There is never a case we need to read more than one extra byte knowing the command, so we should just for loop and pre-emptively grab, we can't really
know the file is bad due to this ahead of
time, just have to play till something else catches!